Add new script to read and clean CSV data
clean_ping_data.py reads CSV data into a pandas DataFrame, cleans the
data, and creates some extra columns that may be useful as features or
for evaluating the success of tutor classification/ranking algorithms.
The modified DataFrame is saved in HDF5 format so it can be quickly
loaded by other scripts.