Add function to create new tutor column
add_weighted_hourly_response computes a weighted average between the
individual response rate per local hour for a tutor and the total
response rate for all tutors. For tutors with many pings, this is
approximately equal to the individual response rate, but if there are
few pings (e.g. for new tutors) it is closer to the overall tutor
response rate. The transition between these limits can be controlled
with an optional weight parameter.